Update: Door module added. Many improvements for CMake building, Removed some "noexcept" keywords (not necessary), some enums replaced by enum classes, and some refactoring
Former-commit-id: 1c25568a4e3bc7291005287675aae3c299f02f83
Former-commit-id: f759d7a8cc9c8359751ebe59a082263ea3875e95

+++ src/dynlib/dynamiclibrary.cpp
@@ -16,7 +16,7 @@ void DynamicLibrary::open(RelocationMode mode)
 {
     char*   err;
 
-    if (!(_handle = dlopen(_file.c_str(), mode)))
+    if (!(_handle = dlopen(_file.c_str(), static_cast<int>(mode))))
     {
         if ((err = dlerror()))
             throw DynLibException(std::string("dlopen(): ") + err);
